+++ pages/index.tsx
@@ -1,4 +1,5 @@
 import Head from 'next/head'
+
 import Button from '../components/Button'
 import Layout from '../components/Layout'
 import styles from './index.module.scss'
@@ -8,7 +9,7 @@ export default function Home(): JSX.Element {
     <Layout>
       <Head>
         <title>Skedge</title>
-        <link rel="icon" href="/favicon.ico" />
+        <link rel="icon" href="https://skedge.pro/favicon.ico" />
       </Head>
 
       <div className={styles.titleContainer} id='title'>