+++ change-log.md
@@ -37,3 +37,8 @@ First major version release (casual though it may be).  All features are conside
 ### Version 1.4.0
 
 - Give Button component the option to target a new tab
+
+### Version 1.5.0
+
+- Add page titles throughout
+- Add Skedge favicon